First pass at PersistentQVM class + app-ng ForestConnection
This commit includes ForestConnection methods to call the new
create-qvm, delete-qvm, qvm-info, and run-program qvm-ng endpoints.

It also includes a PersistentQVM class for wrapping those new
ForestConnection methods.

Additionally, add a separate forest_qvm_ng pytest fixture for
qvm-ng-only tests. This is not yet hooked in to the CI runners and
expects you to have a qvm-ng process running on localhost:5222 or else
the tests will be skipped.